# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/tasks": {
            "post": {
                "description": "接收 JSON body 來建立一個新的待處理任務",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "建立一個新任務",
                "parameters": [
                    {
                        "description": "要建立的任務資訊 (目前僅為範例，未來會簡化)",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.TaskDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "成功建立的任務",
                        "schema": {
                            "$ref": "#/definitions/dto.TaskDTO"
                        }
                    },
                    "400": {
                        "description": "無效的請求",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/tasks/{task_id}": {
            "get": {
                "description": "根據 task_id 取得單一任務的詳細資訊",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "取得一個特定任務",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任務 ID",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功取得的任務",
                        "schema": {
                            "$ref": "#/definitions/dto.TaskDTO"
                        }
                    },
                    "400": {
                        "description": "無效的 ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "找不到任務",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.TaskDTO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "finished_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "playbook": {
                    "description": "例如 \"deploy.yml\"",
                    "type": "string"
                },
                "started_at": {
                    "description": "使用指標(pointer)來允許 null",
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}